Created Web Scraper to Get Gateway Devices
The file gw_device_count.py contains a class GwDeviceCounter that connects to a gateway and returns a count of the associated devices to the user. This supports both HART and ISA devices, and it can retrieve all, live, unreachable, and/or low battery device counts.

This is dependent on the selenium framework and Firefox Gecko web driver. Installation instructions are included at the top of gw_device_count.py.

The other three files are examples of how to connect to different gateways and retrieve statistics of all four types of devices.