Fixed Using Fixed Delays for Getting Device Counts
Previously, the program used a fixed delay before reporting back device counts to allow the Javascript time to update counts when switching tabs. However, this resulted in wasted time where the counts would quickly update, but the program would continue to wait. Additionally, it was possible that the delays given were not large enough at times, so bad results would be returned.

The next system counts the HART and ISA devices and compares it with the total count given by the four different tab types. The program polls these counts, and once the sum of HART and ISA devices are equal, then it knows the Javascript has updated. This allows accurate counts to be consistently reported without wasting excessive amounts of time.