Added Functionality to Return Table as Dictionary
The tables that are displayed on the devices page can be converted to a list of dictionaries where dictionary entries correspond to a cell for each row of the table.

Switch to the desired tab, then call convert_table_into_dicts(device_type = "HART"/"ISA"). The program will go through each page and collect the data. Data is stored as text strings, so any additional processing is done outside this function.