Added Configuration File
To save time when starting a test, the user does not need to manually enter information each time they run the program. When the program is first run, an INI file is created with default configuration parameters. This file can be modified (with one parameter saying to use it) with default hostname, username, etc.