Added HART Tracking
Device and memory usage tracking program not has options to enable/disable HART/ISA tracking. HART devices need to use a web scraper which takes significantly longer than downloading and processing the ISA database. If an alternative, faster method is available in the future to count live HART devices, it should be used in the count_live_hart_devices() function.