Created SSH Helper
Created the start of an SSH helper class that builds upon Paramiko. As needed, this class can be extended in functionality. Currently, its primary feature is to retrieve the memory usage of a given process by passing in the name of the process. The user does NOT need to look up PIDs on their own.

The IsaDeviceCounter class was updated as well to use this helper class instead of the base paramiko SSH client.