Built-in Multithreading Support
If the user wants to decrease the amount of time required to get the memory usage for multiple processes, a tuple or list of processes can be passed to the get_memory_usage_by_multi_name() function to automatically create separate threads for each process request.

The return memory usages are in the order that the processes were passed into the function.