Increased Speed to Check Process Memory Usage and Can Track ISA Data Reliability
Process memory usage on gateways can be checked faster by registering processes with the SSHHelper class. When the register_process() function is called, the process is added to a list within the class, and any associate PIDs are registered with it. Then the check_processes() function can be called which requests memory usage for all registered processes using a single SSH command. This eliminates the need to get the PID and then get the memory usage for each process individually (either sequentially or on separate threads) every time memory usage is checked. The initial process registration can take a noticeable amount of time. However, after registering all processes of interest, getting the memory usage was reduced significantly. For 8 processes, it took about 30 seconds to get them sequentially one at a time, about 15 seconds to do them in parallel on separate threads, and 1 to 2 seconds with this new method.

The IsaDeviceCounter class was also updated to include a new get_device_reliability() function. This looks in the database for the GPDULatency entries under NetworkHealthDevices and calculates the reliability for each devices with the formula reliability = 100 - GPDULatency.