Routing Node is Now a Single Point of Failure
Devices could not join the network in a reasonable amount of time if they were forced to connect to only a single parent node. Advertisements from other devices would cause transmitters to attempt to join and fail for significant amounts of time.

Now, every device except for the specified parent node cannot connect to the gateway. Instead, the parent node connects to the gateway, and then only one other node needs to connect to the parent node. Then every other device can connect with either the parent or any other connected transmitted as long as it is not the gateway.