Changed How Backup Is Downloaded
Instead of using Selenium and Firefox to download backups once they are ready, the requests library is used. Previously, only about 200kB of the backup files would download before closing the browser resulting is many partial backups.

Now, cookies from the Firefox browser session are passed to a requests session. This bypasses the username and password requirement that didn't allow the requests library to be used in the past. A GET request is made to http://<IP or Hostname>/servlet/system_backup.zip to get the most recent backup, and the returned chunks are written into a file.